refactor(frontend): modularize Admin Turtle Match and clarify back navigation
