+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- **Admin Turtle Match navigation**: Removed the header "Back" button next to the page title when a match is selected. The match grid shows a single **Back to upload** control below the header; the detail view only shows **Back to matches**, so reviewers are less likely to leave the flow for the homepage by mistake.
+- **Admin Turtle Match structure**: Refactored the ~1400-line `AdminTurtleMatchPage` into a thin page wrapper, `useAdminTurtleMatch` (state and API handlers), and focused components under `frontend/src/pages/AdminTurtleMatch/` (`MatchGridView`, `MatchDetailView`, `MatchCandidateCard`, `CreateNewTurtleModal`, etc.). Behavior (cross-check, community matches, reference replacement, new-turtle modal) is unchanged.
+
+### Fixed
+
+- **Admin Turtle Match E2E**: Playwright tests now target the **Back to upload** button label instead of the old generic **Back** control.
+
 ## [2.0.7] - 2026-05-16 — Sheets-Browser Null filter count + list-flash fix
 
 ### Fixed
